requirements.txt
PyMuPDF>=1.22.3
Jinja2==3.1.6
num2words==0.5.14
openpyxl==3.1.5
pdfkit==1.0.0
protobuf==6.31.1
python_docx==1.1.2
streamlit==1.45.1
